requirements-deploy.txt
# Production deployment requirements for Prontivus
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Database - PostgreSQL
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication and Security
cryptography==41.0.7
pyotp==2.9.0
qrcode[pil]==7.4.2

# HTTP and API
httpx==0.25.2
requests==2.31.0
aiofiles==23.2.1

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Background tasks
celery==5.3.4
redis==5.0.1

# Monitoring and logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Medical specific
python-dateutil==2.8.2
pytz==2023.3

# PDF Generation and Validation
reportlab==4.0.7
phonenumbers==9.0.14

# Environment management
python-dotenv==1.0.0

# Production server
gunicorn==21.2.0
